requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.9
idna==2.10
multidict==5.1.0
playsound==1.2.2
pycodestyle==2.7.0
pypiwin32==223
python-dotenv==0.16.0
python-vlc==3.0.11115
pyttsx3==2.90
pytube==10.6.1
pywin32==300
requests==2.25.1
toml==0.10.2
twitchio==1.2.2
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==8.1
yarl==1.6.3
